Made ApnsConnectionHandler an inner class of ApnsConnection so some of the gnarlier disconnection methods can be made private.
