+++ src/main/java/com/relayrides/pushy/apns/ApnsConnection.java
@@ -24,6 +24,8 @@ package com.relayrides.pushy.apns;
 import io.netty.bootstrap.Bootstrap;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.SimpleChannelInboundHandler;
 import io.netty.handler.ssl.SslHandler;
 import io.netty.util.concurrent.Future;
 import io.netty.util.concurrent.GenericFutureListener;
@@ -48,6 +50,43 @@ public abstract class ApnsConnection {
 	private volatile boolean closeOnRegistration;
 
 	private static final Logger log = LoggerFactory.getLogger(ApnsConnection.class);
+
+	public static abstract class ApnsConnectionHandler<I> extends SimpleChannelInboundHandler<I> {
+
+		private final ApnsConnection apnsConnection;
+
+		private static final Logger log = LoggerFactory.getLogger(ApnsConnectionHandler.class);
+
+		public ApnsConnectionHandler(final ApnsConnection apnsConnection) {
+			this.apnsConnection = apnsConnection;
+		}
+
+		@Override
+		public void channelRegistered(final ChannelHandlerContext context) throws Exception {
+			super.channelRegistered(context);
+
+			synchronized (this.apnsConnection.getChannelRegistrationMonitor()) {
+				if (this.apnsConnection.shouldCloseOnRegistration()) {
+					log.debug("Channel registered for {}, but shutting down immediately.", this.apnsConnection.getName());
+					context.channel().eventLoop().execute(this.apnsConnection.getImmediateShutdownRunnable());
+				}
+			}
+		}
+
+		@Override
+		public void channelInactive(final ChannelHandlerContext context) throws Exception {
+			super.channelInactive(context);
+
+			// Channel closure implies that the connection attempt had fully succeeded, so we only want to notify
+			// listeners if the handshake has completed. Otherwise, we'll notify listeners of a connection failure (as
+			// opposed to closure) elsewhere.
+			if (this.apnsConnection.hasCompletedHandshake()) {
+				if (this.apnsConnection.getListener() != null) {
+					this.apnsConnection.getListener().handleConnectionClosure(this.apnsConnection);
+				}
+			}
+		}
+	}
 
 	/**
 	 * Constructs a new connection with the given environment and name.
@@ -182,7 +221,7 @@ public abstract class ApnsConnection {
 		}
 	}
 
-	protected Runnable getImmediateShutdownRunnable() {
+	private Runnable getImmediateShutdownRunnable() {
 		final ApnsConnection apnsConnection = this;
 
 		return new Runnable() {
@@ -201,8 +240,12 @@ public abstract class ApnsConnection {
 		};
 	}
 
-	protected boolean shouldCloseOnRegistration() {
+	private boolean shouldCloseOnRegistration() {
 		return this.closeOnRegistration;
+	}
+
+	private Object getChannelRegistrationMonitor() {
+		return this.channelRegistrationMonitor;
 	}
 
 	/**
@@ -219,10 +262,6 @@ public abstract class ApnsConnection {
 			// has not completed.
 			return false;
 		}
-	}
-
-	protected Object getChannelRegistrationMonitor() {
-		return this.channelRegistrationMonitor;
 	}
 
 	/**
